requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
jinja2==3.1.2
weasyprint==62.3
Pillow>=10.0.0
beautifulsoup4>=4.12.0
rembg>=2.0.0
onnxruntime>=1.16.0
openai>=1.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
firebase-admin>=6.0.0

